+++ StudentOglasi/UI/studentoglasi_admin/lib/screens/studetni_list_screen.dart
@@ -63,12 +63,12 @@ class _StudentiListScreenState extends State<StudentiListScreen> {
       'imePrezime': _imePrezimeController.text,
       'fakultetID': selectedFakultet?.id,
       'godinaStudija': selectedGodina,
-      'page': _currentPage + 1, 
+      'page': _currentPage + 1,
       'pageSize': 5,
     });
     setState(() {
       result = data;
-         _totalItems = data.count;
+      _totalItems = data.count;
       int numberPages = calculateNumberPages(_totalItems, 5);
       if (_currentPage >= numberPages) {
         _currentPage = numberPages - 1;
@@ -78,9 +78,9 @@ class _StudentiListScreenState extends State<StudentiListScreen> {
       }
       print(
           "Total items: $_totalItems, Number of pages: $numberPages, Current page after fetch: $_currentPage");
-   
     });
   }
+
   int calculateNumberPages(int totalItems, int pageSize) {
     return (totalItems / pageSize).ceil();
   }
@@ -127,19 +127,23 @@ class _StudentiListScreenState extends State<StudentiListScreen> {
       },
       child: Container(
         child: Column(
-          children: [_buildSearch(), _buildDataListView(),if(_currentPage>=0 && numberPages-1>=_currentPage)
-           CustomPaginator(
-                      numberPages: numberPages,
-                      initialPage: _currentPage,
-                      onPageChange: (int index) {
-                        setState(() {
-                          _currentPage = index;
-                          _fetchData();
-                        });
-                      },
-                      pageController: _pageController,
-                      fetchData: _fetchData,
-                    ),],
+          children: [
+            _buildSearch(),
+            _buildDataListView(),
+            if (_currentPage >= 0 && numberPages - 1 >= _currentPage)
+              CustomPaginator(
+                numberPages: numberPages,
+                initialPage: _currentPage,
+                onPageChange: (int index) {
+                  setState(() {
+                    _currentPage = index;
+                    _fetchData();
+                  });
+                },
+                pageController: _pageController,
+                fetchData: _fetchData,
+              ),
+          ],
         ),
       ),
     );
@@ -347,8 +351,51 @@ class _StudentiListScreenState extends State<StudentiListScreen> {
                                       color: Colors.red,
                                     ),
                                     onPressed: () async {
-                                      await _studentiProvider.delete(e.id);
-                                      await _fetchData();
+                                      bool confirmDelete = await showDialog(
+                                        context: context,
+                                        builder: (BuildContext context) {
+                                          return AlertDialog(
+                                            title: Row(
+                                              mainAxisAlignment:
+                                                  MainAxisAlignment
+                                                      .spaceBetween,
+                                              children: [
+                                                Text("Potvrda brisanja"),
+                                                IconButton(
+                                                  icon: Icon(Icons.close),
+                                                  onPressed: () {
+                                                    Navigator.of(context).pop(
+                                                        false); 
+                                                  },
+                                                ),
+                                              ],
+                                            ),
+                                            content: Text(
+                                                "Da li ste sigurni da želite izbrisati?"),
+                                            actions: [
+                                              TextButton(
+                                                onPressed: () {
+                                                  Navigator.of(context).pop(
+                                                      false); 
+                                                },
+                                                child: Text("Ne"),
+                                              ),
+                                              TextButton(
+                                                onPressed: () {
+                                                  Navigator.of(context).pop(
+                                                      true);  
+                                                },
+                                                child: Text("Da"),
+                                              ),
+                                            ],
+                                          );
+                                        },
+                                      );
+
+                                      if (confirmDelete == true) {
+                                        await _studentiProvider.delete(e.id);
+                                        await _fetchData();
+                                      }
                                     },
                                   ),
                                 ],
